requirements.txt
streamlit
langchain
langchain-core
langchain-groq
langchain-community
langgraph
faiss-cpu
python-dotenv